splitting profile and addresses tests files
